llvm: Pass number of inputs to 'evaluate' function
Commit b8dffcc8050df6cfab840e0a0731d5118f4c6f72
	("llvm/OptimizationControlMechanism: Move construction of
	simulation input out of "evaluate" function")
and commit aa63fc706f9a31eb188a54d9e1a90a47a0e4bd4c
	("llvm, gird_evaluate: Use 'state_feature_values' to construct input")

changed compiled evaluate to take inputs in the same format as
composition run. The input is then directly passed to simulation runs.
However, the compiled codepath was still hardcoding the number of inputs
to 1.
This change fixes that and passes the number of inputs to 'evaluate'
along with the inputs for simulations.

CUDA also needs to account for the number of input sets when uploading
parameters to shared memory.

Signed-off-by: Jan Vesely <jan.vesely@rutgers.edu>